+++ lib/ui/groups/GroupMemberNotificationsPanel.dart
@@ -113,6 +113,7 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
   Widget _buildNotificationsContent() {
     List<Widget> preferenceWidgets = [];
     MemberNotificationsPreferences? memberPreferences = _member?.notificationsPreferences;
+    preferenceWidgets.add(_buildGlobalNotificationsInfo());
 
     preferenceWidgets.add(_EnabledToggleButton(
         enabled: _toggleButtonEnabled,
@@ -171,6 +172,64 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
     ]));
 
     return Container(child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: preferenceWidgets));
+  }
+
+  _buildGlobalNotificationsInfo(){
+    // bool groupNotificationsEnabled = FirebaseMessaging().notifyGroupUpdates == true;
+    bool groupPostNotificationsEnabled = FirebaseMessaging().notifyGroupPostUpdates == true;
+    bool groupEventsNotificationsEnabled = FirebaseMessaging().notifyGroupEventsUpdates == true;
+    bool groupInvitationsNotificationsEnabled = FirebaseMessaging().notifyGroupInvitationsUpdates == true;
+    bool groupPollsNotificationsEnabled = FirebaseMessaging().notifyGroupPollsUpdates == true;
+
+    List<Widget> widgets = [];
+    widgets.add(Row(children: [
+        Expanded(
+          child: Container(
+              padding: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
+              color: Colors.white,
+              child: Text(Localization().getStringEx('panel.group_member_notifications.member.info.heading.msg', 'Global Group Notification Settings'), //TBD localize
+                  style: Styles().textStyles?.getTextStyle("widget.title.regular")),
+          )
+        )
+    ],));
+
+    // widgets.add(ToggleRibbonButton(
+    //     borderRadius: BorderRadius.zero,
+    //     label: Localization().getStringEx("panel.settings.notifications.group_updates", "Group Updates"),
+    //     toggled: groupNotificationsEnabled,
+    //     textStyle: groupNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled") : Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.fat.disabled")
+    // ));
+    widgets.add(Row(children: [Expanded(child: Container(color: Styles().colors!.white, child: Padding(padding: EdgeInsets.only(left: 10), child: Column(children: [
+      ToggleRibbonButton(
+          toggled: groupPostNotificationsEnabled,
+          borderRadius: BorderRadius.zero,
+          label: Localization().getStringEx("panel.settings.notifications.group_updates.posts.label", "Posts"),
+          textStyle: groupPostNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
+      ),
+      ToggleRibbonButton(
+          toggled: groupEventsNotificationsEnabled,
+          borderRadius: BorderRadius.zero,
+          label: Localization().getStringEx("panel.settings.notifications.group_updates.event.label", "Event"),
+          textStyle: groupEventsNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
+      ),
+      ToggleRibbonButton(
+          toggled: groupInvitationsNotificationsEnabled,
+          borderRadius: BorderRadius.zero,
+          label: Localization().getStringEx("panel.settings.notifications.group_updates.invitations.label", "Invitations"),
+          textStyle: groupInvitationsNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
+      ),
+      ToggleRibbonButton(
+          toggled:  groupPollsNotificationsEnabled,
+          borderRadius: BorderRadius.zero,
+          label: Localization().getStringEx("panel.settings.notifications.group_updates.polls.label", "Polls"),
+          textStyle: groupPollsNotificationsEnabled ? Styles().textStyles?.getTextStyle("panel.group_member_notifications.toggle_button.title.small.enabled"): Styles().textStyles?.getTextStyle("panel.settings.toggle_button.title.small.disabled")
+      )
+    ]))))]));
+    widgets.add(Container(color:Styles().colors!.surfaceAccent,height: 1,));
+
+      return Container(
+        child: Padding(padding: EdgeInsets.all(0), child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: widgets)),
+      );
   }
 
   void _checkNotificationsEnabled() {