+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Research Project updates [#2533](https://github.com/rokwire/illinois-app/issues/2533).
 - Make the MTD route lines thicker [#2536](https://github.com/rokwire/illinois-app/issues/2536).
 
+###Added
+- Show global group notification setting in Group by Group panel [#2538](https://github.com/rokwire/illinois-app/issues/2538).
+
 ## [4.2.19] - 2022-11-15
 ### Added
 - Initialize research confirmation flag in Group Create/Settings panels, require user consent before requesting to join research project that requires confirmation [#2531](https://github.com/rokwire/illinois-app/issues/2531).